Supporting the ProxyCommand feature
Now, instead of just ignoring the ProxyCommand var, the ssh library is able to retrieve
this configuration from the `~/.ssh/conf` file and use it to proxy the connection between
the client and the SSH server.

This commit introduces the ssh.proxy module that implements the interface required by
the client.Transport class. But instead of using sockets, this class writes and reads info
from a peopened program that actually talks with the ssh server.